move BasicStore to a sample file and rename to SampleDB
